requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
opencv-python
sorobn
graphviz
gym
seaborn
ipython
jupyter